+++ src/Model/Table/VideosTable.php
@@ -2,23 +2,13 @@
 /**
  * This file is part of me-cms-youtube.
  *
- * me-cms-youtube is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
  *
- * me-cms-youtube is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with me-cms-youtube.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
- * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
- * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
- * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms-youtube
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
  */
 namespace MeCmsYoutube\Model\Table;
 
@@ -73,14 +63,14 @@ class VideosTable extends AppTable
 
     /**
      * Internal method to get information about a video
-     * @param string $id Video ID
+     * @param string $videoId Video ID
      * @return mixed Object or `false`
      * @uses $Youtube
      * @uses MeCmsYoutube\Utility\Youtube::getInfo()
      */
-    protected function _getInfo($id)
+    protected function getInfo($videoId)
     {
-        return $this->Youtube->getInfo($id);
+        return $this->Youtube->getInfo($videoId);
     }
 
     /**
@@ -124,12 +114,12 @@ class VideosTable extends AppTable
      * @param \Cake\Datasource\EntityInterface $entity Entity
      * @param \ArrayObject $options Options
      * @return bool
-     * @uses _getInfo()
+     * @uses getInfo()
      */
     public function beforeSave(Event $event, EntityInterface $entity, ArrayObject $options)
     {
         if ((empty($entity->seconds) || empty($entity->duration)) && !empty($entity->youtube_id)) {
-            $info = $this->_getInfo($entity->youtube_id);
+            $info = $this->getInfo($entity->youtube_id);
 
             if (empty($entity->seconds)) {
                 $entity->seconds = $info->seconds;
@@ -201,7 +191,7 @@ class VideosTable extends AppTable
      * @return Query $query Query object
      * @uses $cache
      */
-    protected function _getRandomSpots()
+    protected function getRandomSpotsQuery()
     {
         return $this->find()
             ->select('youtube_id')
@@ -217,11 +207,11 @@ class VideosTable extends AppTable
      * Gets random spots
      * @param int $limit Limit
      * @return \Cake\Collection\Collection Collection
-     * @uses _getRandomSpots()
+     * @uses getRandomSpotsQuery()
      */
     public function getRandomSpots($limit = 1)
     {
-        return $this->_getRandomSpots()->sample($limit);
+        return $this->getRandomSpotsQuery()->sample($limit);
     }
 
     /**