+++ tests/TestCase/View/Helper/MenuHelperTest.php
@@ -22,9 +22,9 @@
  */
 namespace MeCmsYoutube\Test\TestCase\Utility;
 
-use Cake\TestSuite\TestCase;
 use Cake\View\View;
 use MeCmsYoutube\View\Helper\MenuHelper;
+use MeTools\TestSuite\TestCase;
 
 /**
  * YoutubeTest class
@@ -47,17 +47,6 @@ class MenuHelperTest extends TestCase
         parent::setUp();
 
         $this->Menu = new MenuHelper(new View);
-    }
-
-    /**
-     * Teardown any static object changes and restore them
-     * @return void
-     */
-    public function tearDown()
-    {
-        parent::tearDown();
-
-        unset($this->Menu);
     }
 
     /**