+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 2.x branch
 ## 2.10 branch
+### 2.10.5
+* added initial schema of the plugin database. Removed `UpdateShell` class;
+* the MIT license has been applied;
+* significantly improved all tests.
+
 ### 2.10.4
 * fixed bug for widgets: they do not show anything if there are no records.